Modal Component 코드 중복 제거
- naming 변경
- form이 포함된 modal들을 `DataEditableModal.jsx`로 추상화
