+++ .babelrc
@@ -1,4 +1,9 @@
 {
+  "plugins": [
+    ["babel-plugin-transform-builtin-extend", {
+      "globals": ["Error", "Array"]
+    }]
+  ],
   "env": {
     "test": { "presets": [["env", { "targets": { "node": "current" } }]] },
     "development": {
