+++ package.json
@@ -38,6 +38,7 @@
   "dependencies": {
     "axios": "^0.16.2",
     "babel-loader": "^7.1.1",
+    "babel-plugin-transform-builtin-extend": "^1.1.2",
     "babel-polyfill": "^6.20.0",
     "babel-preset-react": "^6.16.0",
     "babel-register": "^6.18.0",
